exercise 4.21 completed

+++ controllers/blogs.js
@@ -12,10 +12,6 @@ blogsRouter.post("/", async (request, response) => {
     const body = request.body
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-    if(!decodedToken.id) {
-        return response.status(401).json({ error: "invalid token" })
-    }
-
     const user = await User.findById(decodedToken.id)
 
     const blog = new Blog({
@@ -40,8 +36,19 @@ blogsRouter.get("/:id", async (request, response) => {
 
 blogsRouter.delete("/:id", async (request, response) => {
     const id = request.params.id
-    await Blog.findByIdAndDelete(id)
-    response.status(204).end()
+    const blog = await Blog.findById(id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!blog) {
+        return response.status(404).json({ error: "blog not found" })
+    }
+
+    if (blog.user.toString() === decodedToken.id.toString()) {
+        await Blog.findByIdAndDelete(id)
+        response.status(204).end()
+    } else {
+        response.status(401).json({ error: "unauthorized" })
+    }
 })
 
 blogsRouter.put("/:id", async (request, response) => {
